Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "googlemaps>=4.10.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.4",
    "litellm>=1.74.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
